requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
xgboost==3.2.0
